refactor(combobox): removed combobox controlled mode for now
