+++ packages/components/combobox/src/Combobox.doc.mdx
@@ -91,14 +91,6 @@ AutoSuggest is used as the default behaviour. The user can type anything in the 
 
 <Canvas of={stories.Default} />
 
-### Controlled
-
-<Canvas of={stories.Controlled} />
-
-### Controlled open state
-
-<Canvas of={stories.ControlledOpenState} />
-
 ### Custom filtering
 
 Disable `autoFilter` to implement your own logic to filter out items depending on the inputValue or some external logic.
@@ -175,10 +167,6 @@ In multiple selection mode, the input will go back to empty state after each sel
 This is up to the developer to make it clear to the user which items are selected, by using other components such as chips, for example.
 
 <Canvas of={stories.MultipleSelection} />
-
-### Controlled
-
-<Canvas of={stories.MultipleSelectionControlled} />
 
 ## Advanced usage
