+++ packages/components/combobox/src/tests/singleSelection.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { useState } from 'react'
+// import { useState } from 'react'
 import { describe, expect, it } from 'vitest'
 
 import { Combobox } from '..'
@@ -95,43 +95,43 @@ describe('Combobox', () => {
       expect(screen.getByLabelText('selected')).toBeVisible()
     })
 
-    it('should control value', async () => {
-      const user = userEvent.setup()
+    // it('should control value', async () => {
+    //   const user = userEvent.setup()
 
-      // Given we control value by outside state and selected value
-      const ControlledImplementation = () => {
-        const [value, setValue] = useState('book-1')
-        const [inputValue, setInputValue] = useState('')
+    //   // Given we control value by outside state and selected value
+    //   const ControlledImplementation = () => {
+    //     const [value, setValue] = useState('book-1')
+    //     const [inputValue, setInputValue] = useState('')
 
-        return (
-          <Combobox value={value} onValueChange={setValue}>
-            <Combobox.Trigger>
-              <Combobox.Input aria-label="Book" value={inputValue} onValueChange={setInputValue} />
-            </Combobox.Trigger>
-            <Combobox.Popover>
-              <Combobox.Items>
-                <Combobox.Item value="book-1">War and Peace</Combobox.Item>
-                <Combobox.Item value="book-2">1984</Combobox.Item>
-                <Combobox.Item value="book-3">Pride and Prejudice</Combobox.Item>
-              </Combobox.Items>
-            </Combobox.Popover>
-          </Combobox>
-        )
-      }
+    //     return (
+    //       <Combobox value={value} onValueChange={setValue}>
+    //         <Combobox.Trigger>
+    //           <Combobox.Input aria-label="Book" value={inputValue} onValueChange={setInputValue} />
+    //         </Combobox.Trigger>
+    //         <Combobox.Popover>
+    //           <Combobox.Items>
+    //             <Combobox.Item value="book-1">War and Peace</Combobox.Item>
+    //             <Combobox.Item value="book-2">1984</Combobox.Item>
+    //             <Combobox.Item value="book-3">Pride and Prejudice</Combobox.Item>
+    //           </Combobox.Items>
+    //         </Combobox.Popover>
+    //       </Combobox>
+    //     )
+    //   }
 
-      render(<ControlledImplementation />)
+    //   render(<ControlledImplementation />)
 
-      expect(getItem('War and Peace')).toHaveAttribute('aria-selected', 'true')
+    //   expect(getItem('War and Peace')).toHaveAttribute('aria-selected', 'true')
 
-      expect(screen.getByDisplayValue('')).toBeInTheDocument()
+    //   expect(screen.getByDisplayValue('')).toBeInTheDocument()
 
-      // when the user select another item
-      await user.click(getInput('Book'))
-      await user.click(getItem('Pride and Prejudice'))
+    //   // when the user select another item
+    //   await user.click(getInput('Book'))
+    //   await user.click(getItem('Pride and Prejudice'))
 
-      // Then the selected value has been updated
-      expect(screen.getByDisplayValue('Pride and Prejudice')).toBeInTheDocument()
-    })
+    //   // Then the selected value has been updated
+    //   expect(screen.getByDisplayValue('Pride and Prejudice')).toBeInTheDocument()
+    // })
 
     it('should select item using autoFilter (keyboard)', async () => {
       const user = userEvent.setup()
